Pull up PropagateEmptyRelation
